+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # ChangeLog
 
-* merge 1.20 changes
+* added the ability for any model to use LDLib CTM in the textures
+  * works on both forge and fabric